+++ statsd.go
@@ -21,6 +21,7 @@ import (
 	"fmt"
 	"io"
 	"math/rand"
+	"os"
 	"strconv"
 	"strings"
 	"sync"
@@ -33,6 +34,7 @@ const (
 
 type StatsClient interface {
 	SetPrefix(prefix string)
+	Unique(stat string, count int, rate float64) error
 	Increment(stat string, count int, rate float64) error
 	Decrement(stat string, count int, rate float64) error
 	Duration(stat string, duration time.Duration, rate float64) error
@@ -99,6 +101,20 @@ func (c *client) SetPrefix(prefix string) {
 	c.m.Lock()
 	defer c.m.Unlock()
 	c.prefix = strings.TrimRight(prefix, ".") + "."
+}
+
+// makeStatsPrefix will create a stats key prefix based on the given environment, application name, and hostname.
+func MakeStatsdPrefix(namespace, app, hostname string) string {
+	underscoreHostname := strings.Replace(hostname, ".", "_", -1)
+	return fmt.Sprintf("%s.%s.%s.", namespace, app, underscoreHostname)
+}
+
+func GetHostname() (string, error) {
+	hostname, err := os.Hostname()
+	if err != nil {
+		hostname = "UNKNOWN"
+	}
+	return hostname, err
 }
 
 // Increment the counter for the given bucket.